feat: replace `AddLinkHeaderPostProcessor` with `AddLinkHeaderResponseInterceptor`
update unit tests to use appropriate `ISpecification`
