+++ src/Agenda.API/Features/Appointments/v1/Search/SearchAppointmentsEndpoint.cs
@@ -23,6 +23,7 @@ public class SearchAppointmentsEndpoint : Endpoint<SearchAppointmentRequest, Ok<
     private readonly IUnitOfWorkFactory _unitOfWorkFactory;
     private readonly LinkGenerator _linkGenerator;
     private readonly CurrentRequestMetadataInfoProvider _currentRequestMetadataInfo;
+    private readonly ILoggerFactory _loggerFactory;
 
     /// <summary>
     /// Builds a new <see cref="SearchAppointmentsEndpoint"/> instance
@@ -30,13 +31,17 @@ public class SearchAppointmentsEndpoint : Endpoint<SearchAppointmentRequest, Ok<
     /// <param name="unitOfWorkFactory">Gives access to the underlying datastore</param>
     /// <param name="linkGenerator">Helper to generate links between resources.</param>
     /// <param name="currentRequestMetadataInfo"></param>
+    /// <param name="loggerFactory"></param>
+
     public SearchAppointmentsEndpoint(IUnitOfWorkFactory unitOfWorkFactory,
                                       LinkGenerator linkGenerator,
-                                      CurrentRequestMetadataInfoProvider currentRequestMetadataInfo)
+                                      CurrentRequestMetadataInfoProvider currentRequestMetadataInfo,
+                                      ILoggerFactory loggerFactory)
     {
         _unitOfWorkFactory = unitOfWorkFactory;
         _linkGenerator = linkGenerator;
         _currentRequestMetadataInfo = currentRequestMetadataInfo;
+        _loggerFactory = loggerFactory;
     }
 
     /// <inheritdoc />
@@ -45,7 +50,7 @@ public class SearchAppointmentsEndpoint : Endpoint<SearchAppointmentRequest, Ok<
         Verbs(Http.GET, Http.HEAD);
         Routes("/appointments");
         AllowAnonymous();
-        PostProcessor<AddLinkHeaderPostProcessor>();
+        ResponseInterceptor(new AddLinkHeaderResponseInterceptor(_loggerFactory.CreateLogger<AddLinkHeaderResponseInterceptor>()));
     }
 
     /// <inheritdoc />
