+++ tests/Agenda.API.UnitTests/Features/Appointments/v1/Delete/DeleteByIdEndpointShould.cs
@@ -131,7 +131,7 @@ public class DeleteByIdEndpointShould
         IReadOnlyList<Appointment> appointments = appointmentsInStore.Value;
 
         A.CallTo(() => _fakeRepository.Any(An<IFilterSpecification<Appointment>>._, A<CancellationToken>._))
-            .ReturnsLazily((Expression<Func<Appointment, bool>> predicate, CancellationToken _) => appointments.Any(predicate.Compile()));
+            .ReturnsLazily((IFilterSpecification<Appointment> predicate, CancellationToken _) => appointments.Any(predicate.Filter.Compile()));
 
         A.CallTo(() => _fakeRepository.Delete(An<IFilterSpecification<Appointment>>._, A<CancellationToken>._))
             .DoesNothing();
